Fix “Back” button in Booking Request Interface
Initially, when the "Back" button is clicked, the system will direct to the View Item interface, which is as same as when the system will direct if the user clicked "Submit" button. 

The changes that has been made is to change the "Back" button href link to the main page instead of view item interface.